🐞 Fix: 박자 정확도 저장 오류 문제 해결 - 박자 정확도 저장 위치 수정 - 현재 음정 바 위치를 프로바이더에서 알 수 있게 수정
